+++ package.json
@@ -19,17 +19,18 @@
     "dayjs": "^1.11.5",
     "leaflet": "^1.8.0",
     "next": "12.2.5",
+    "next-intl": "^2.7.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-leaflet": "^4.0.2",
     "sass": "^1.54.5"
   },
   "devDependencies": {
-    "@types/react-leaflet": "^2.8.2",
     "@types/leaflet": "^1.7.11",
     "@types/node": "18.7.4",
     "@types/react": "18.0.17",
     "@types/react-dom": "18.0.6",
+    "@types/react-leaflet": "^2.8.2",
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "eslint": "8.22.0",
     "eslint-config-next": "12.2.5",
