+++ src/components/main-page/Index.tsx
@@ -1,7 +1,7 @@
+import Head from "next/head";
 import { Hero } from "./Hero/Hero";
 import { EventDetails } from "./EventDetails/EventDetails";
 import { Registration } from "./Registration/Registration";
-import Head from "next/head";
 
 export function Index() {
   return (
