+++ src/pages/_app.tsx
@@ -1,16 +1,18 @@
 import type { AppProps } from "next/app";
-import { LanguageProvider } from "translations/context/LanguageContext";
 import { Footer, Navigation } from "components/ui";
 import "styles/globals.css";
 import "styles/variables.css";
+import { NextIntlProvider } from "next-intl";
 
 function App({ Component, pageProps }: AppProps) {
+  const { locale, translations } = pageProps.localization;
+
   return (
-    <LanguageProvider localization={pageProps.localization}>
+    <NextIntlProvider messages={translations} locale={locale}>
       <Navigation />
       <Component {...pageProps} />
       <Footer />
-    </LanguageProvider>
+    </NextIntlProvider>
   );
 }
