+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { GetStaticProps, NextPage } from "next";
 import { Localization } from "translations/types";
 import { Index } from "components/main-page/Index";
-import { getLocalizationProps } from "translations/context/LanguageContext";
+import { getLocalizationProps } from "translations";
 
 const IndexPage: NextPage<{
   localization: Localization;
