+++ src/translations/locales/pl.json
